+++ composeApp/src/commonMain/kotlin/com/lfssolutions/retialtouch/domain/model/productBarCode/BarcodeDao.kt
@@ -1,6 +1,6 @@
 package com.lfssolutions.retialtouch.domain.model.productBarCode
 
 data class BarcodeDao(
-    val barcodeId: Int=0,
+    val barcodeId: Long=0,
     val barcode: Barcode= Barcode(),
 )
